+++ generators/voxel_generator.cpp
@@ -6,8 +6,6 @@ VoxelGenerator::VoxelGenerator() {
 
 void VoxelGenerator::generate_block(VoxelBlockRequest &input) {
 	ERR_FAIL_COND(input.voxel_buffer.is_null());
-	try_call_script(this, VoxelStringNames::get_singleton()->generate_block,
-			input.voxel_buffer, input.origin_in_voxels.to_vec3(), input.lod, nullptr);
 }
 
 //bool VoxelGenerator::is_thread_safe() const {
@@ -30,7 +28,6 @@ void VoxelGenerator::_b_generate_block(Ref<VoxelBuffer> out_buffer, Vector3 orig
 }
 
 void VoxelGenerator::_bind_methods() {
-	// Note: C++ inheriting classes don't need to re-bind these, because they are bindings that call the actual virtual methods
-
-	ClassDB::bind_method(D_METHOD("generate_block", "out_buffer", "origin_in_voxels", "lod"), &VoxelGenerator::_b_generate_block);
+	ClassDB::bind_method(D_METHOD("generate_block", "out_buffer", "origin_in_voxels", "lod"),
+			&VoxelGenerator::_b_generate_block);
 }