+++ streams/voxel_stream.cpp
@@ -7,12 +7,10 @@ VoxelStream::VoxelStream() {
 
 void VoxelStream::emerge_block(Ref<VoxelBuffer> out_buffer, Vector3i origin_in_voxels, int lod) {
 	ERR_FAIL_COND(out_buffer.is_null());
-	try_call_script(this, VoxelStringNames::get_singleton()->emerge_block, out_buffer, origin_in_voxels.to_vec3(), lod, nullptr);
 }
 
 void VoxelStream::immerge_block(Ref<VoxelBuffer> buffer, Vector3i origin_in_voxels, int lod) {
 	ERR_FAIL_COND(buffer.is_null());
-	try_call_script(this, VoxelStringNames::get_singleton()->immerge_block, buffer, origin_in_voxels.to_vec3(), lod, nullptr);
 }
 
 void VoxelStream::emerge_blocks(Vector<VoxelBlockRequest> &p_blocks) {
@@ -49,10 +47,6 @@ void VoxelStream::_immerge_block(Ref<VoxelBuffer> buffer, Vector3 origin_in_voxe
 }
 
 int VoxelStream::get_used_channels_mask() const {
-	Variant ret;
-	if (try_call_script(this, VoxelStringNames::get_singleton()->get_used_channels_mask, nullptr, 0, &ret)) {
-		return ret;
-	}
 	return 0;
 }
 
@@ -70,8 +64,6 @@ bool VoxelStream::has_script() const {
 }
 
 void VoxelStream::_bind_methods() {
-	// TODO Make these proper virtual, it confuses C# bindings
-	// Note: C++ inheriting classes don't need to re-bind these, because they are bindings that call the actual virtual methods
 	ClassDB::bind_method(D_METHOD("emerge_block", "out_buffer", "origin_in_voxels", "lod"), &VoxelStream::_emerge_block);
 	ClassDB::bind_method(D_METHOD("immerge_block", "buffer", "origin_in_voxels", "lod"), &VoxelStream::_immerge_block);
 	ClassDB::bind_method(D_METHOD("get_used_channels_mask"), &VoxelStream::_get_used_channels_mask);