+++ util/utility.cpp
@@ -26,6 +26,7 @@ bool is_mesh_empty(Ref<Mesh> mesh_ref) {
 
 bool try_call_script(const Object *obj, StringName method_name, const Variant **args, unsigned int argc, Variant *out_ret) {
 	ScriptInstance *script = obj->get_script_instance();
+	// TODO Is has_method() needed? I've seen `call()` being called anyways in ButtonBase
 	if (script == nullptr || !script->has_method(method_name)) {
 		return false;
 	}