+++ voxel_string_names.cpp
@@ -14,10 +14,11 @@ void VoxelStringNames::destroy_singleton() {
 }
 
 VoxelStringNames::VoxelStringNames() {
-	emerge_block = StaticCString::create("emerge_block");
-	immerge_block = StaticCString::create("immerge_block");
-	generate_block = StaticCString::create("generate_block");
-	get_used_channels_mask = StaticCString::create("get_used_channels_mask");
+	_emerge_block = StaticCString::create("_emerge_block");
+	_immerge_block = StaticCString::create("_immerge_block");
+	_generate_block = StaticCString::create("_generate_block");
+	_get_used_channels_mask = StaticCString::create("_get_used_channels_mask");
+
 	block_loaded = StaticCString::create("block_loaded");
 	block_unloaded = StaticCString::create("block_unloaded");