+++ voxel_string_names.h
@@ -17,10 +17,11 @@ public:
 
 	VoxelStringNames();
 
-	StringName emerge_block;
-	StringName immerge_block;
-	StringName generate_block;
-	StringName get_used_channels_mask;
+	StringName _emerge_block;
+	StringName _immerge_block;
+	StringName _generate_block;
+	StringName _get_used_channels_mask;
+
 	StringName block_loaded;
 	StringName block_unloaded;